refactor: WSEditorCommandGroupContent to function-based
